requirements.txt
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
asyncpg==0.29.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
docker==7.1.0
filelock==3.16.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
jiter==0.5.0
openai==1.44.0
packaging==24.1
pydantic==2.9.0
pydantic_core==2.23.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
